chore: add instructions on routeHelpNameMap
